UBUNTU: [Packaging] Drop unreachable config annotations
In the Jammy environment, some config options cannot be enabled as
enforced in the parent kernel. Add code to automatically adjust those.
This was attempted to be manual edits to the annotations. However those
get overwritten on each rebase. So instead add it to the local-mangle
script but run only if the compiler set to be used is gcc-11.

Ignore: yes
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>